App creates some events and tickets + an admin when it is run

+++ src/main/java/ca/mcgill/ecse428/ESCAPE/dto/TicketRequestDto.java
@@ -7,7 +7,7 @@ public class TicketRequestDto {
 	private double price;
 	private int eventId;
 	
-	public TicketRequestDto(String name, int price, int eventId) {
+	public TicketRequestDto(String name, double price, int eventId) {
 		this.name = name;
 		this.price = price;
 		this.eventId = eventId;
@@ -28,7 +28,7 @@ public class TicketRequestDto {
 	public double getPrice() {
 		return price;
 	}
-	public void setPrice(int price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
